requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-storages==1.13.2
boto3==1.34.0
gunicorn==21.2.0  
